+++ init/01_grant_global_access.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+mariadb -u root -p"$MYSQL_ROOT_PASSWORD" <<-EOSQL
+    GRANT ALL PRIVILEGES ON *.* TO '$MYSQL_USER'@'%';
+    FLUSH PRIVILEGES;
+EOSQL