+++ init/01_user-permissions.sql
@@ -1,2 +1,0 @@
--- Give the hypermedia user all privileges on all databases
-GRANT ALL PRIVILEGES ON *.* TO 'hypermedia'@'%';